Validate link types in imports in the link parser
This commit adds a validation check to make sure that the link types
parsed from html match strings that we expect.

To build that list, I iterated over a random `course` instance's methods
using (`course.methods`), and checked each to see if
`course.respond_to?(method)` and
`course.send(method).respond_to?(:scope)` was true (This is how it was
intended to be used, and only these can have meaningful behavior).
From there, I removed ones that seemed obviously not supported.

I'm sure more might be able to be removed.

Also added to that list (which are not methods on course instances)
were:

```
media_attachments_iframe
modules
pages
wiki
```

test plan:

export a course with a link to a wiki page, and manually
modify that link from `$WIKI_REFERENCE$/pages/this-is-a-page?wrap=1`
to something like: `$WIKI_REFERENCE$/foo/this-is-a-page?wrap=1`. The
import should not fail, and the link's href should not be changed.

flag=none

Change-Id: I2243fc4da2e1cd87b52b206bb857e2ec75d518d7
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/320881
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: James Logan <james.logan@instructure.com>
QA-Review: James Logan <james.logan@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>